libhsmd: Migrate handle_sign_bolt12
